Fix: cast COUNT result to int in unique validator to prevent false positives
PDO can return COUNT(*) as a string, causing strict === 0 comparison to fail
and incorrectly reporting emails as already taken on both create and update.

+++ core/Validator.php
@@ -392,7 +392,7 @@ class Validator
                 );
             }
             
-            return ($result['count'] ?? 0) === 0;
+            return (int) ($result['count'] ?? 0) === 0;
         } catch (\Exception $e) {
             // If database check fails, allow (don't block user)
             error_log("Unique validation failed: " . $e->getMessage());
